Extract inlay hint walker to a separate file
